Added cpp file for callback functions
The file contains also const variables and functions used by the
aforementioned callbacks
